some fixes

+++ src/main/java/cn/nukkit/entity/EntityLiving.java
@@ -56,15 +56,15 @@ public abstract class EntityLiving extends Entity implements EntityDamageable {
         super.initEntity();
 
         if (this.namedTag.contains("HealF")) {
-            this.namedTag.putShort("Health", this.namedTag.getShort("HealF"));
+            this.namedTag.putFloat("Health", this.namedTag.getShort("HealF"));
             this.namedTag.remove("HealF");
         }
 
         if (!this.namedTag.contains("Health") || !(this.namedTag.get("Health") instanceof ShortTag)) {
-            this.namedTag.putShort("Health", this.getMaxHealth());
+            this.namedTag.putFloat("Health", this.getMaxHealth());
         }
 
-        this.setHealth(this.namedTag.getShort("Health"));
+        this.setHealth(this.namedTag.getFloat("Health"));
     }
 
     @Override
@@ -82,7 +82,7 @@ public abstract class EntityLiving extends Entity implements EntityDamageable {
     @Override
     public void saveNBT() {
         super.saveNBT();
-        this.namedTag.putShort("Health", (int) this.getHealth());
+        this.namedTag.putFloat("Health", this.getHealth());
     }
 
     public boolean hasLineOfSight(Entity entity) {
